Flesh out some initial thoughts on the characteristics of a model
Including an example to help tinker with the user interface for modeling a
process for a mock heroku-based web application.

What's missing right now for sure is the concept of credentials and how those
might fit into the world. That will require more thought.

The syntax is at the moment heavily inspired by Declarative Pipeline from
Jenkins, but not actually compatible.