Support configuring the channels directly from our configuration
This required the concepts of stateful and stateless channels to be introduced,
which will have different persistence guarantees.

The two types of channels will help the eventbus determine where it can be cheap
and efficient, and where it should involve more expensive storage. When I think
of an eventbus interface implemented on top of Redis, I would use BLPOP lists
for stateful channels and pub/sub for stateless channels.